Support single_user_mode with Ansible 2.9
