--- plugins/cache/__init__.py
@@ -1,55 +1,0 @@
-#
-# (c) 2020 Red Hat Inc.
-#
-# This file is part of Ansible
-#
-# Ansible is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# Ansible is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with Ansible.  If not, see <http://www.gnu.org/licenses/>.
-#
-
-from __future__ import absolute_import, division, print_function
-
-__metaclass__ = type
-
-from abc import abstractmethod
-from ansible.plugins import AnsiblePlugin
-
-
-class NetworkBaseCacheModule(AnsiblePlugin):
-    def __init__(self, *args, **kwargs):
-        pass
-
-    @abstractmethod
-    def get(self, key):
-        pass
-
-    @abstractmethod
-    def set(self, key, value):
-        pass
-
-    @abstractmethod
-    def keys(self):
-        pass
-
-    @abstractmethod
-    def flush(self):
-        pass
-
-    def lookup(self, key):
-        return self.get(key)
-
-    def populate(self, key, value):
-        self.set(key, value)
-
-    def invalidate(self):
-        self.flush()
