+++ plugins/cache/base.py
@@ -21,33 +21,35 @@ from __future__ import absolute_import, division, print_function
 
 __metaclass__ = type
 
-DOCUMENTATION = """
-    cache: memory
-    short_description: RAM backed, non persistent cache.
-    description:
-        - RAM backed cache that is not persistent.
-        - Tailored for networking use case.
-    version_added: 2.0.0
-    author: Ansible Networking Team
-"""
-
-from ansible_collections.ansible.netcommon.plugins.cache.base import (
-    NetworkBaseCacheModule,
-)
+from abc import abstractmethod
+from ansible.plugins import AnsiblePlugin
 
 
-class CacheModule(NetworkBaseCacheModule):
+class NetworkBaseCacheModule(AnsiblePlugin):
     def __init__(self, *args, **kwargs):
-        self._cache = {}
+        pass
 
+    @abstractmethod
     def get(self, key):
-        return self._cache.get(key)
+        pass
 
+    @abstractmethod
     def set(self, key, value):
-        self._cache[key] = value
+        pass
 
+    @abstractmethod
     def keys(self):
-        return self._cache.keys()
+        pass
 
+    @abstractmethod
     def flush(self):
-        self._cache = {}
+        pass
+
+    def lookup(self, key):
+        return self.get(key)
+
+    def populate(self, key, value):
+        self.set(key, value)
+
+    def invalidate(self):
+        self.flush()
